GraalVM tests renamed
